Ensure that the server owner is given ID priority and clean code
The server owner is now given ID (e.g. sorting) priority so that
if another member has the same ID as the owner, the owner's ID
won't be the one changed. This is important as bots cannot change
the server owners ID.

The missing default configuration value for the server ID was
added. Additionally, the code was cleaned up via the additon of comments
and variable renaming.