Only add IDs to active players
The bot now maintains a list of active players and will only give
IDs to players who are active. A new function to get a list of
active players was created. Additionaly the adjust nicknames
function now takes a reference to the bot class as apposed to
several individual arguments.

A configuration field for the game start date was added.